GH-1163: Moved update password component.
