+++ projects/storefrontlib/src/lib/my-account/update-password/components/update-password/update-password.component.html
@@ -1,0 +1,72 @@
+<div class="col-md-6 justify-content-center container">
+  <form
+    (submit)="resetPassword()"
+    [formGroup]="form"
+    class="cx-update-password-component "
+  >
+    <div class="form-group">
+      <label>
+        <span class="label-content">New Password</span>
+        <input
+          class="form-control"
+          [class.is-invalid]="
+            form.get('password').invalid &&
+            (submited ||
+              (form.get('password').dirty && form.get('password').touched))
+          "
+          type="password"
+          name="password"
+          placeholder="Password"
+          formControlName="password"
+        />
+        <div
+          class="invalid-feedback"
+          *ngIf="
+            form.get('password').invalid &&
+            (submited ||
+              (form.get('password').dirty && form.get('password').touched))
+          "
+        >
+          <span
+            >Password must be six characters minimum, with one uppercase letter,
+            one number, one symbol</span
+          >
+        </div>
+      </label>
+    </div>
+
+    <div class="form-group">
+      <label>
+        <span class="label-content">Confirm New Password</span>
+        <input
+          class="form-control"
+          [class.is-invalid]="
+            form.hasError('NotEqual') &&
+            (submited ||
+              (form.get('repassword').dirty && form.get('repassword').touched))
+          "
+          type="password"
+          name="confirmpassword"
+          placeholder="Confirm Password"
+          formControlName="repassword"
+        />
+        <div
+          class="invalid-feedback"
+          *ngIf="
+            form.hasError('NotEqual') &&
+            (submited ||
+              (form.get('repassword').dirty && form.get('repassword').touched))
+          "
+        >
+          <span>Both password must match</span>
+        </div>
+      </label>
+    </div>
+
+    <div class="form-group">
+      <button class="btn btn-block btn-primary" type="submit">
+        Reset Password
+      </button>
+    </div>
+  </form>
+</div>
